Using snapshot sdk for parsing metadata
