+++ MapboxSearch/gradle/versions.gradle
@@ -45,7 +45,7 @@ ext {
     common_sdk_version = '24.5.0-beta.4'
     mapbox_base_version = '0.8.0'
 
-    search_native_version = '2.2.0-beta.1'
+    search_native_version = '2.2.0-sbs_metadata_children-SNAPSHOT'
 
     detekt_version = '1.19.0'
